backend/requirements.txt
# ── Core ──
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
pydantic==2.10.4
pydantic-settings==2.7.1

# ── Task Queue ──
# celery[redis]==5.4.0
# redis==5.2.1

# ── Database ──
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.14.1
greenlet==3.1.1

# ── Log Parsing ──
py3grok==0.2.0
regex

# ── PII Sanitization ──
presidio-analyzer==2.2.357
presidio-anonymizer==2.2.357
spacy==3.8.3

# ── LLM Providers ──
google-genai==1.3.0
groq==0.15.0

# ── HTTP Client ──
httpx==0.28.1

# ── Auth ──
python-jose[cryptography]==3.3.0

# ── Utilities ──
orjson==3.10.13
